Rename main theme
