+++ src/client/Theme/index.ts
@@ -1,2 +1,2 @@
-export { default as lightTheme } from './LightTheme';
+export { default as mainTheme } from './MainTheme';
 export { default as darkTheme } from './DarkTheme';
